Fix colors, layout and scrolling behavior. (#294)
* feat!: integrate cafe into mensa view

* fix!: handle missing mensa data.

* fix!: fix auto-scrolling on slow devices.

Introduce requestAnimationFrame to improve animation handling. Instead of just drawing, use requestAnimationFrame which informs the browser that an animation will be taking place.

This commit also introduces a fix where tiles that are refetched, will continue to scroll. This was previously not possible because they updated the DOM and thus removed any scrolling events. Now using a MutationObserver we can look for reloaded tiles and initialize scrolling.

Also RIP Prinz Karl.

* fix: fix colors and update tiles layout.